requirements.txt
flask
flask_pymongo
werkzeug
google-cloud-vision
gpt4all
gunicorn
dotenv